Move telemetry command to preferences.contribution
